requirements.txt
appdata==2.1.2
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.0
click==8.0.3
colorama==0.4.5
colored==1.4.3
coverage==6.4.1
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
Django==4.0.5
django-debug-toolbar==3.5.0
django-static-fontawesome==5.15.4.0
entrypoints==0.4
executing==0.8.3
flake8==4.0.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
nest-asyncio==1.5.5
outcome==1.2.0
packaging==21.3
parameterized==0.8.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pywin32==304
pyzmq==23.2.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.3.0
to-requirements.txt==1.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
wsproto==1.1.0
